Exercises 4.15.-4.23. done

+++ Osa4/bloglist/controllers/blogs.js
@@ -1,8 +1,11 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
 
@@ -17,20 +20,42 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
+    const user = request.user
+    
+    if (!user) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
 
-    const blog = new Blog ({
+    const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes
+        likes: body.likes,
+        user: user._id,
     })
+    
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const user = request.user
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    
+    if (!(token && decodedToken.id)) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() === user.id.toString() ) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(401).end()
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -43,7 +68,11 @@ blogsRouter.put('/:id', async (request, response) => {
         likes: body.likes
     }
 
-    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new: true})
+    const updatedBlog = await Blog.findByIdAndUpdate(
+        request.params.id,
+        blog,
+        { new: true }
+    )
     response.json(updatedBlog)
 })
